restructuring of the project
