+++ src/blocks/rating/styles.less
