+++ src/pages/afisha-page.jsx
@@ -1,0 +1,10 @@
+import {PageHeader} from '../blocks/page-header'
+import { Afisha } from '../blocks/afisha'
+
+export function AfishaPage() {
+    return(
+    <>
+        <PageHeader/>
+        <Afisha/>
+    </>)
+}
